reviewing again the EventCoutner stuff... some work is still needed

+++ drools-framework-runtime-base/src/main/java/org/chtijbug/drools/runtime/RuleBasePackage.java
@@ -35,7 +35,7 @@ public interface RuleBasePackage {
 
     public HistoryListener getHistoryListener();
 
-    public int getRuleBaseID();
+    public Long getRuleBaseID();
 
     public void dispose();
