+++ drools-framework-runtime-base/src/main/java/org/chtijbug/drools/runtime/RuleBaseSession.java
@@ -75,6 +75,6 @@ public interface RuleBaseSession {
 
     public int getSessionId();
 
-    public int getRuleBaseID();
+    public Long getRuleBaseID();
 
 }
