+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/session/SessionEvent.java
@@ -27,8 +27,7 @@ import java.util.Date;
  */
 public class SessionEvent extends HistoryEvent {
 
-    public SessionEvent(Long eventID, int ruleBaseId, int sessionId) {
-
+    public SessionEvent(Long eventID, Long ruleBaseId, int sessionId) {
         super(eventID, new Date(), TypeEvent.Session);
         this.setRuleBaseID(ruleBaseId);
         this.setSessionId(sessionId);
